backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
channels==4.1.0
channels-redis==4.2.0
uvicorn==0.30.1
redis==5.0.1
pymongo[srv]==4.8.0
dnspython==2.6.1
gunicorn==21.2.0
